+++ src/store/nodes.ts
@@ -11,6 +11,7 @@ import {
 
 type NodesState = {
   filter: string
+  nodeRoles: string[]
   hideAttributesRegex: string
 } & PaginationStorePartial &
   ReloadIntervalStorePartial
@@ -21,12 +22,13 @@ export const useNodesStore = () => {
   return defineStore(`nodes-${clusterUuid}`, {
     state: (): NodesState => ({
       filter: '',
+      nodeRoles: [],
       hideAttributesRegex: DEFAULT_HIDE_NODE_ATTRIBUTES_REGEX,
       reloadInterval: null,
       pagination: paginationStoreDefaultProps('name')
     }),
     persist: {
-      pick: ['filter', 'hideAttributesRegex', ...persistReloadIntervalProps(), ...persistPaginationProps()],
+      pick: ['filter', 'nodeRoles', 'hideAttributesRegex', ...persistReloadIntervalProps(), ...persistPaginationProps()],
       key: `nodes-${clusterUuid}`
     }
   })()