+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * show search aggregations, fixes [#247](https://github.com/cars10/elasticvue/issues/247)
 * can filter indices by alias, fixes [#337](https://github.com/cars10/elasticvue/issues/337)
+* can filter node roles, fixes [#336](https://github.com/cars10/elasticvue/issues/336)
 * i18n: adds traditional chinese, thanks @PeterDaveHello
 * i18n: adds korean, thanks @webee0317